refactor: move activities into folders
