+++ app/src/main/java/vn/edu/usth/mobile_app/ui/modeldetails/similarmodels/MoreSimilarModelsViewModel.kt
@@ -1,4 +1,4 @@
-package vn.edu.usth.mobile_app.ui.modeldetails
+package vn.edu.usth.mobile_app.ui.modeldetails.similarmodels
 
 import androidx.lifecycle.ViewModel
 import vn.edu.usth.mobile_app.model.ModelData
